Move avro C++ code to  tensorflow_io/core/kernels/avro, and clang-format fix
Directory under `tensorflow_io/core/` is reserved for language level (python/swift) directory so move
move avro C++ code to  tensorflow_io/core/kernels/avro
