+++ tensorflow_io/core/BUILD
@@ -277,20 +277,20 @@ cc_library(
 cc_library(
     name = "avro_ops",
     srcs = [
-        "avro/avro_dataset_kernels.cc",
-        "avro/avro_record_dataset_kernels.cc",
-        "avro/parse_avro_kernels.cc",
+        "kernels/avro/avro_dataset_kernels.cc",
+        "kernels/avro/avro_record_dataset_kernels.cc",
+        "kernels/avro/parse_avro_kernels.cc",
         "kernels/avro_kernels.cc",
         "ops/avro_ops.cc",
     ],
     hdrs = [
-        "avro/avro_record_dataset_kernels.h",
+        "kernels/avro/avro_record_dataset_kernels.h",
     ],
     copts = tf_io_copts(),
     linkstatic = True,
     deps = [
         "//tensorflow_io/core:dataset_ops",
-        "//tensorflow_io/core/avro/utils:avro_utils",
+        "//tensorflow_io/core/kernels/avro/utils:avro_utils",
     ],
     alwayslink = 1,
 )
