+++ tensorflow_io/core/kernels/avro/utils/BUILD
