+++ tensorflow_io/core/kernels/avro/utils/avro_file_stream_reader.h
@@ -18,7 +18,7 @@ limitations under the License.
 #include "tensorflow/core/lib/io/buffered_inputstream.h"
 #include "tensorflow/core/lib/io/inputbuffer.h"
 #include "tensorflow/core/lib/io/random_inputstream.h"
-#include "tensorflow_io/core/avro/utils/avro_parser_tree.h"
+#include "tensorflow_io/core/kernels/avro/utils/avro_parser_tree.h"
 
 namespace tensorflow {
 namespace data {
