+++ tensorflow_io/core/kernels/avro/utils/name_utils.cc
@@ -9,7 +9,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
-#include "tensorflow_io/core/avro/utils/name_utils.h"
+#include "tensorflow_io/core/kernels/avro/utils/name_utils.h"
 
 #include "absl/strings/str_join.h"
