+++ tensorflow_io/core/kernels/avro/utils/prefix_tree.cc
@@ -9,7 +9,8 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
-#include "tensorflow_io/core/avro/utils/prefix_tree.h"
+#include "tensorflow_io/core/kernels/avro/utils/prefix_tree.h"
+
 #include <sstream>
 
 namespace tensorflow {
