+++ tensorflow_io/core/kernels/avro/utils/value_buffer.cc
@@ -9,7 +9,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
-#include "tensorflow_io/core/avro/utils/value_buffer.h"
+#include "tensorflow_io/core/kernels/avro/utils/value_buffer.h"
 
 namespace tensorflow {
 namespace data {
